Move pending diffs into new patches module
