+++ automerge-backend/src/patches.rs
@@ -1,0 +1,5 @@
+mod pending_diff;
+mod patch_workshop;
+
+pub(crate) use pending_diff::{Edits, PendingDiff, PendingDiffs};
+pub(crate) use patch_workshop::PatchWorkshop;
